# artifacts/api-server/src/routes/patients.ts
/**
 * patients.ts — Express router for patient CRUD endpoints.
 *
 * Routes:
 *   GET    /api/patients          — list all admitted patients
 *   POST   /api/patients          — admit a new patient
 *   GET    /api/patients/:id      — fetch a single patient record
 *   PATCH  /api/patients/:id      — update patient info (ward, condition, etc.)
 *   DELETE /api/patients/:id      — discharge & remove patient
 *
 * All database access goes through the shared Drizzle ORM `db` instance
 * exported from @workspace/db.  Zod schemas generated by Orval are used
 * for request body validation to keep the server-side contract in sync
 * with the OpenAPI spec.
 */
import { Router } from "express";
import { eq } from "drizzle-orm";
import { db, patientsTable } from "@workspace/db";
import {
  CreatePatientBody,
  UpdatePatientBody,
  GetPatientParams,
  UpdatePatientParams,
  DeletePatientParams,
} from "@workspace/api-zod";

const router = Router();

// ─── GET /patients ─────────────────────────────────────────────────────────────
router.get("/patients", async (req, res) => {
  try {
    // Retrieve all patients ordered by admission date descending (most recent first)
    const patients = await db
      .select()
      .from(patientsTable)
      .orderBy(patientsTable.admittedAt);

    return res.json(patients);
  } catch (err) {
    req.log.error({ err }, "Failed to list patients");
    return res.status(500).json({ error: "Internal server error" });
  }
});

// ─── POST /patients ────────────────────────────────────────────────────────────
router.post("/patients", async (req, res) => {
  // Validate request body against the Orval-generated Zod schema
  const parseResult = CreatePatientBody.safeParse(req.body);
  if (!parseResult.success) {
    return res.status(400).json({ error: parseResult.error.format() });
  }

  try {
    const [patient] = await db
      .insert(patientsTable)
      .values({
        name: parseResult.data.name,
        age: parseResult.data.age,
        gender: parseResult.data.gender,
        ward: parseResult.data.ward,
        bed: parseResult.data.bed,
        condition: parseResult.data.condition,
        diagnosis: parseResult.data.diagnosis ?? null,
      })
      .returning();

    return res.status(201).json(patient);
  } catch (err) {
    req.log.error({ err }, "Failed to create patient");
    return res.status(500).json({ error: "Internal server error" });
  }
});

// ─── GET /patients/:id ────────────────────────────────────────────────────────
router.get("/patients/:id", async (req, res) => {
  // Validate path params
  const paramResult = GetPatientParams.safeParse({ id: Number(req.params.id) });
  if (!paramResult.success) {
    return res.status(400).json({ error: "Invalid patient ID" });
  }

  try {
    const [patient] = await db
      .select()
      .from(patientsTable)
      .where(eq(patientsTable.id, paramResult.data.id));

    if (!patient) {
      return res.status(404).json({ error: "Patient not found" });
    }

    return res.json(patient);
  } catch (err) {
    req.log.error({ err }, "Failed to get patient");
    return res.status(500).json({ error: "Internal server error" });
  }
});

// ─── PATCH /patients/:id ──────────────────────────────────────────────────────
router.patch("/patients/:id", async (req, res) => {
  const paramResult = UpdatePatientParams.safeParse({ id: Number(req.params.id) });
  const bodyResult = UpdatePatientBody.safeParse(req.body);

  if (!paramResult.success || !bodyResult.success) {
    return res.status(400).json({ error: "Invalid request" });
  }

  try {
    const [updated] = await db
      .update(patientsTable)
      .set({
        ...bodyResult.data,
        updatedAt: new Date(),
      })
      .where(eq(patientsTable.id, paramResult.data.id))
      .returning();

    if (!updated) {
      return res.status(404).json({ error: "Patient not found" });
    }

    return res.json(updated);
  } catch (err) {
    req.log.error({ err }, "Failed to update patient");
    return res.status(500).json({ error: "Internal server error" });
  }
});

// ─── DELETE /patients/:id ─────────────────────────────────────────────────────
router.delete("/patients/:id", async (req, res) => {
  const paramResult = DeletePatientParams.safeParse({ id: Number(req.params.id) });
  if (!paramResult.success) {
    return res.status(400).json({ error: "Invalid patient ID" });
  }

  try {
    const deleted = await db
      .delete(patientsTable)
      .where(eq(patientsTable.id, paramResult.data.id))
      .returning();

    if (deleted.length === 0) {
      return res.status(404).json({ error: "Patient not found" });
    }

    // 204 No Content — successful deletion with no body
    return res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Failed to delete patient");
    return res.status(500).json({ error: "Internal server error" });
  }
});

export default router;